Refactor the template loader to allow for includes.
- The 'base' templates are moved to the hooks/base folder.
- The Jinja2 environment will process expressions that start
  with //jinja:
- Hook template will go through a final compile stage to ensure
  that all Jinja2 tokens have been processed before being sent
  off to Frida for loading.
- The jailbreak/root detection and simulation hooks now include a
  small template for the `common_path` variable so that in the
  future only this file needs updating with more paths.
