+++ objection/hooks/android/root/_common_paths.js
@@ -1,0 +1,19 @@
+// Contains common paths used in both the root
+// disable and simulation hooks. These paths are included
+// using the //jinja: directive
+
+var common_paths = [
+    '/data/local/bin/su',
+    '/data/local/su',
+    '/data/local/xbin/su',
+    '/dev/com.koushikdutta.superuser.daemon/',
+    '/sbin/su',
+    '/system/app/Superuser.apk',
+    '/system/bin/failsafe/su',
+    '/system/bin/su',
+    '/system/etc/init.d/99SuperSUDaemon',
+    '/system/sd/xbin/su',
+    '/system/xbin/busybox',
+    '/system/xbin/daemonsu',
+    '/system/xbin/su',
+];
