+++ objection/hooks/base/generic-base.js
