+++ objection/hooks/ios/jailbreak/_common_paths.js
@@ -1,0 +1,36 @@
+// Contains common paths used in both the jailbreak
+// disable and simulation hooks. These paths are included
+// using the //jinja: directive
+
+var common_paths = [
+    '/Applications/Cydia.app',
+    '/Applications/FakeCarrier.app',
+    '/Applications/Icy.app',
+    '/Applications/IntelliScreen.app',
+    '/Applications/MxTube.app',
+    '/Applications/RockApp.app',
+    '/Applications/SBSetttings.app',
+    '/Applications/WinterBoard.app',
+    '/Applications/blackra1n.app',
+    '/Library/MobileSubstrate/DynamicLibraries/Veency.plist',
+    '/Library/MobileSubstrate/MobileSubstrate.dylib',
+    '/System/Library/LaunchDaemons/com.ikey.bbot.plist',
+    '/System/Library/LaunchDaemons/com.saurik.Cy@dia.Startup.plist',
+    '/bin/bash',
+    '/bin/sh',
+    '/etc/apt',
+    '/etc/ssh/sshd_config',
+    '/private/var/stash',
+    '/private/var/tmp/cydia.log',
+    '/usr/bin/cycript',
+    '/usr/bin/ssh',
+    '/usr/bin/sshd',
+    '/usr/libexec/sftp-server',
+    '/usr/libexec/sftp-server',
+    '/usr/libexec/ssh-keysign',
+    '/usr/sbin/sshd',
+    '/var/cache/apt',
+    '/var/lib/cydia',
+    '/var/log/syslog',
+    '/var/tmp/cydia.log',
+];
