+++ objection/hooks/ios/jailbreak/simulate.js
@@ -1,38 +1,8 @@
 // Attempts to simulate a Jailbroken device by responding to
 // common types of checks positively.
 
-var common_paths = [
-    '/bin/sh',
-    '/etc/apt',
-    '/bin/bash',
-    '/usr/bin/ssh',
-    '/usr/bin/sshd',
-    '/var/cache/apt',
-    '/usr/sbin/sshd',
-    '/var/lib/cydia',
-    '/var/log/syslog',
-    '/usr/bin/cycript',
-    '/var/tmp/cydia.log',
-    '/private/var/stash',
-    '/etc/ssh/sshd_config',
-    '/Applications/Icy.app',
-    '/Applications/Cydia.app',
-    '/usr/libexec/ssh-keysign',
-    '/usr/libexec/sftp-server',
-    '/Applications/MxTube.app',
-    '/usr/libexec/sftp-server',
-    '/Applications/RockApp.app',
-    '/private/var/tmp/cydia.log',
-    '/Applications/blackra1n.app',
-    '/Applications/FakeCarrier.app',
-    '/Applications/SBSetttings.app',
-    '/Applications/WinterBoard.app',
-    '/Applications/IntelliScreen.app',
-    '/Library/MobileSubstrate/MobileSubstrate.dylib',
-    '/System/Library/LaunchDaemons/com.ikey.bbot.plist',
-    '/Library/MobileSubstrate/DynamicLibraries/Veency.plist',
-    '/System/Library/LaunchDaemons/com.saurik.Cy@dia.Startup.plist',
-];
+// Get the common_paths for iOS
+//jinja: include 'ios/jailbreak/_common_paths.js'
 
 // Hook NSFileManager calls and check if it is to a common path.
 // TODO: Hook fopen too.
